backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers
djangorestframework
pytz==2024.1